+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityComboBoxPanel.java
@@ -111,6 +111,18 @@ public final class EntityComboBoxPanel extends JPanel {
 		Builder includeEditButton(boolean includeEditButton);
 
 		/**
+		 * @param confirmAdd true if adding an item should be confirmed
+		 * @return this builder instance
+		 */
+		Builder confirmAdd(boolean confirmAdd);
+
+		/**
+		 * @param confirmEdit true if editing an item should be confirmed
+		 * @return this builder instance
+		 */
+		Builder confirmEdit(boolean confirmEdit);
+
+		/**
 		 * Default false
 		 * @param buttonsFocusable true if the buttons should be focusable
 		 * @return this builder instance
@@ -175,6 +187,18 @@ public final class EntityComboBoxPanel extends JPanel {
 		@Override
 		public Builder includeEditButton(boolean includeEditButton) {
 			this.includeEditButton = includeEditButton;
+			return this;
+		}
+
+		@Override
+		public Builder confirmAdd(boolean confirmAdd) {
+			this.entityComboBoxBuilder.confirmAdd(confirmAdd);
+			return this;
+		}
+
+		@Override
+		public Builder confirmEdit(boolean confirmEdit) {
+			this.entityComboBoxBuilder.confirmEdit(confirmEdit);
 			return this;
 		}